Split sdk in a different package
