+++ sdk/vrealize_sdk_test.go
@@ -1,4 +1,4 @@
-package vrealize
+package sdk
 
 // import (
 // 	"errors"
