+++ utils/schema.go
@@ -1,10 +1,7 @@
-package vrealize
+package utils
 
 import (
 	"time"
-
-	"github.com/hashicorp/terraform/helper/schema"
-	"github.com/vmware/terraform-provider-vra7/utils"
 )
 
 // ResourceActionTemplate - is used to store information
@@ -193,85 +190,4 @@ type CatalogRequest struct {
 		ID    string `json:"id"`
 		Label string `json:"label"`
 	} `json:"catalogItemRef"`
-}
-
-//ResourceMachine - use to set resource fields
-func ResourceMachine() *schema.Resource {
-	return &schema.Resource{
-		Create: createResource,
-		Read:   readResource,
-		Update: updateResource,
-		Delete: deleteResource,
-		Schema: resourceSchema(),
-	}
-}
-
-//set_resource_schema - This function is used to update the catalog item template/blueprint
-//and replace the values with user defined values added in .tf file.
-func resourceSchema() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		utils.CatalogName: {
-			Type:     schema.TypeString,
-			Optional: true,
-		},
-		utils.CatalogID: {
-			Type:     schema.TypeString,
-			Computed: true,
-			Optional: true,
-		},
-		utils.BusinessGroupID: {
-			Type:     schema.TypeString,
-			Computed: true,
-			Optional: true,
-		},
-		utils.BusinessGroupName: {
-			Type:     schema.TypeString,
-			Computed: true,
-			Optional: true,
-		},
-		utils.WaitTimeout: {
-			Type:     schema.TypeInt,
-			Optional: true,
-			Default:  15,
-		},
-		utils.RequestStatus: {
-			Type:     schema.TypeString,
-			Computed: true,
-			ForceNew: true,
-		},
-		utils.FailedMessage: {
-			Type:     schema.TypeString,
-			Computed: true,
-			ForceNew: true,
-			Optional: true,
-		},
-		utils.DeploymentConfiguration: {
-			Type:     schema.TypeMap,
-			Optional: true,
-			Elem: &schema.Schema{
-				Type:     schema.TypeMap,
-				Optional: true,
-				Elem:     schema.TypeString,
-			},
-		},
-		utils.ResourceConfiguration: {
-			Type:     schema.TypeMap,
-			Optional: true,
-			Computed: true,
-			Elem: &schema.Schema{
-				Type:     schema.TypeMap,
-				Optional: true,
-				Elem:     schema.TypeString,
-			},
-		},
-		utils.CatalogConfiguration: {
-			Type:     schema.TypeMap,
-			Optional: true,
-			Elem: &schema.Schema{
-				Type:     schema.TypeMap,
-				Optional: true,
-				Elem:     schema.TypeString,
-			},
-		},
-	}
 }
